+++ recipe/flow_framework.php
@@ -42,6 +42,7 @@ task('deploy:publish_resources', function () {
 task('deploy', [
     'deploy:info',
     'deploy:prepare',
+    'deploy:lock',
     'deploy:release',
     'deploy:update_code',
     'deploy:vendors',
@@ -50,6 +51,7 @@ task('deploy', [
     'deploy:run_migrations',
     'deploy:publish_resources',
     'deploy:symlink',
+    'deploy:unlock',
     'cleanup',
 ])->desc('Deploy your project');