+++ recipe/wordpress.php
@@ -20,7 +20,6 @@ task('deploy', [
     'deploy:release',
     'deploy:update_code',
     'deploy:shared',
-    'deploy:vendors',
     'deploy:writable',
     'deploy:symlink',
     'deploy:unlock',