+++ src/Collection/PersistentCollection.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 /* (c) Anton Medvedev <anton@medv.io>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -9,9 +9,6 @@ namespace Deployer\Collection;
 
 class PersistentCollection extends Collection
 {
-    /**
-     * @var string
-     */
     private $file;
 
     public function __construct(string $file, array $collection = [])