+++ src/Executor/ExecutorInterface.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 /* (c) Anton Medvedev <anton@medv.io>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -16,5 +16,5 @@ interface ExecutorInterface
      * @param Task[] $tasks
      * @param Host[] $hosts
      */
-    public function run($tasks, $hosts);
+    public function run(array $tasks, array $hosts);
 }