+++ src/Executor/SeriesExecutor.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 /* (c) Anton Medvedev <anton@medv.io>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -31,23 +31,20 @@ class SeriesExecutor implements ExecutorInterface
      */
     private $informer;
 
-    /**
-     * @param InputInterface $input
-     * @param OutputInterface $output
-     * @param Informer $informer
-     */
-    public function __construct(InputInterface $input, OutputInterface $output, Informer $informer)
-    {
+    public function __construct(
+        InputInterface $input,
+        OutputInterface $output,
+        Informer $informer
+    ) {
         $this->input = $input;
         $this->output = $output;
         $this->informer = $informer;
     }
 
-
     /**
      * {@inheritdoc}
      */
-    public function run($tasks, $hosts)
+    public function run(array $tasks, array $hosts)
     {
         $localhost = new Localhost();
         foreach ($tasks as $task) {